refactor: put header logic inside header and its components
